+++ prisma/migrations/20230803174805_altered_items/migration.sql
@@ -1,5 +1,0 @@
--- AlterTable
-ALTER TABLE "users" ADD COLUMN     "admin" BOOLEAN NOT NULL DEFAULT false;
-
--- AlterTable
-ALTER TABLE "usersClient" ADD COLUMN     "admin" BOOLEAN NOT NULL DEFAULT false;