+++ prisma/migrations/20230809185116_created_data/migration.sql
@@ -1,13 +1,16 @@
 -- CreateTable
-CREATE TABLE "usersClient" (
+CREATE TABLE "users" (
     "id" TEXT NOT NULL,
     "name" TEXT NOT NULL,
     "email" TEXT NOT NULL,
     "password" TEXT NOT NULL,
+    "contact" TEXT NOT NULL,
+    "local" TEXT NOT NULL,
+    "admin" BOOLEAN NOT NULL DEFAULT false,
     "created_at" TIMESTAMP(3) DEFAULT CURRENT_TIMESTAMP,
     "updated_at" TIMESTAMP(3) DEFAULT CURRENT_TIMESTAMP,
 
-    CONSTRAINT "usersClient_pkey" PRIMARY KEY ("id")
+    CONSTRAINT "users_pkey" PRIMARY KEY ("id")
 );
 
 -- CreateTable
@@ -24,8 +27,14 @@ CREATE TABLE "categories" (
 CREATE TABLE "veicules" (
     "id" TEXT NOT NULL,
     "name" TEXT NOT NULL,
+    "model" TEXT NOT NULL,
     "price" TEXT NOT NULL,
     "description" TEXT NOT NULL,
+    "km" TEXT NOT NULL,
+    "march" TEXT NOT NULL,
+    "year" TEXT NOT NULL,
+    "color" TEXT NOT NULL,
+    "city" TEXT NOT NULL,
     "banner" TEXT NOT NULL,
     "created_at" TIMESTAMP(3) DEFAULT CURRENT_TIMESTAMP,
     "updated_at" TIMESTAMP(3) DEFAULT CURRENT_TIMESTAMP,
@@ -39,9 +48,10 @@ CREATE TABLE "orders" (
     "id" TEXT NOT NULL,
     "status" BOOLEAN NOT NULL DEFAULT false,
     "draft" BOOLEAN NOT NULL DEFAULT true,
-    "name" TEXT NOT NULL,
+    "name" TEXT,
     "created_at" TIMESTAMP(3) DEFAULT CURRENT_TIMESTAMP,
     "updated_at" TIMESTAMP(3) DEFAULT CURRENT_TIMESTAMP,
+    "user_id" TEXT NOT NULL,
 
     CONSTRAINT "orders_pkey" PRIMARY KEY ("id")
 );
@@ -60,6 +70,9 @@ CREATE TABLE "items" (
 
 -- AddForeignKey
 ALTER TABLE "veicules" ADD CONSTRAINT "veicules_category_id_fkey" FOREIGN KEY ("category_id") REFERENCES "categories"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
+
+-- AddForeignKey
+ALTER TABLE "orders" ADD CONSTRAINT "orders_user_id_fkey" FOREIGN KEY ("user_id") REFERENCES "users"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
 
 -- AddForeignKey
 ALTER TABLE "items" ADD CONSTRAINT "items_order_id_fkey" FOREIGN KEY ("order_id") REFERENCES "orders"("id") ON DELETE RESTRICT ON UPDATE CASCADE;