+++ src/services/user/DetailUserService.ts
@@ -4,7 +4,7 @@ import prismaClient from "../../prisma";
 class DetailUserService{
     async execute(user_id: string){
 
-        const user = await prismaClient.userAdmin.findFirst({
+        const user = await prismaClient.users.findFirst({
             where:{
                 id: user_id
             },