+++ CHANGELOG.md
@@ -32,6 +32,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Multer
 - configured banner upload
 - configured cadaster of vehicles
+- Created list by category
+
 
 ### Fixed